pyproject.toml
[tool.poetry]
name = "filmer"
version = "0.1.0"
description = "film info"
authors = ["eduzen <eduardo.a.enriquez@gmail.com>"]

[tool.pytest.ini_options]
addopts = "-s"
DJANGO_SETTINGS_MODULE = "filmer.settings"


[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
pendulum = "^2.1.2"
django-extensions = "^3.1.3"
django-crispy-forms = "^1.12.0"
django-debug-toolbar = "^3.2.2"
whitenoise = "^5.3.0"
Markdown = "^3.3.4"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
sentry-sdk = "^1.3.1"
django-health-check = "^3.16.4"
django-filter = "^2.4.0"
requests = "^2.26.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.1"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
pytest-django = "^4.4.0"
richer = "^0.1.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
ipdb = "^0.13.9"
factory-boy = "^3.2.0"
pytest-sugar = "^0.9.4"
pytest_async = "^0.1.1"
wdb = "^3.3.0"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py39']
